+++ plugins/modules/dellemc_configure_idrac_eventing.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 #
-# Dell EMC OpenManage Ansible Modules
-# Version 6.0.0
+# Dell OpenManage Ansible Modules
+# Version 7.0.0
 # Copyright (C) 2018-2022 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
@@ -88,7 +88,7 @@ options:
         choices: [Enabled, Disabled]
     smtp_ip_address:
         type: str
-        description: SMTP IP address for communication.
+        description: Enter the IPv4 or IPv6 address of the SMTP server or the FQDN or DNS name.
     smtp_port:
         type: str
         description: SMTP Port number for access.
@@ -104,7 +104,7 @@ requirements:
 author: "Felix Stephen (@felixs88)"
 notes:
     - This module requires 'Administrator' privilege for I(idrac_user).
-    - Run this module from a system that has direct access to Dell EMC iDRAC.
+    - Run this module from a system that has direct access to Dell iDRAC.
     - This module supports C(check_mode).
 """