+++ plugins/modules/idrac_lifecycle_controller_status_info.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 #
-# Dell EMC OpenManage Ansible Modules
-# Version 5.0.1
+# Dell OpenManage Ansible Modules
+# Version 7.0.0
 # Copyright (C) 2018-2022 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
@@ -19,7 +19,7 @@ module: idrac_lifecycle_controller_status_info
 short_description: Get the status of the Lifecycle Controller
 version_added: "2.1.0"
 description:
-    - This module shows the status of the Lifecycle Controller on a Dell EMC PowerEdge server.
+    - This module shows the status of the Lifecycle Controller on a Dell PowerEdge server.
 extends_documentation_fragment:
     - dellemc.openmanage.idrac_auth_options
 
@@ -30,7 +30,7 @@ author:
     - "Rajeev Arakkal (@rajeevarakkal)"
     - "Anooja Vardhineni (@anooja-vardhineni)"
 notes:
-    - Run this module from a system that has direct access to DellEMC iDRAC.
+    - Run this module from a system that has direct access to Dell iDRAC.
     - This module supports C(check_mode).
 """
 
@@ -52,7 +52,7 @@ msg:
   type: str
   sample: "Successfully fetched the lifecycle controller status."
 lc_status_info:
-  description: Displays the status of the Lifecycle Controller on a Dell EMC PowerEdge server.
+  description: Displays the status of the Lifecycle Controller on a Dell PowerEdge server.
   returned: success
   type: dict
   sample: {