+++ plugins/modules/ome_firmware.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 #
-# Dell EMC OpenManage Ansible Modules
-# Version 5.0.1
+# Dell OpenManage Ansible Modules
+# Version 7.0.0
 # Copyright (C) 2019-2022 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
@@ -60,8 +60,8 @@ options:
       - This option is case sensitive.
       - This is applicable to I(device_service_tag), I(device_id), and I(baseline_name).
     type: list
-    elements: str
     default: []
+    elements: str
   devices:
     description:
       - This option allows to select components on each device for firmware update.
@@ -101,7 +101,7 @@ author:
     - "Felix Stephen (@felixs88)"
     - "Jagadeesh N V (@jagadeeshnv)"
 notes:
-    - Run this module from a system that has direct access to Dell EMC OpenManage Enterprise.
+    - Run this module from a system that has direct access to Dell OpenManage Enterprise.
     - This module supports C(check_mode).
 '''