+++ plugins/modules/ome_firmware_baseline.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 #
-# Dell EMC OpenManage Ansible Modules
-# Version 5.0.1
+# Dell OpenManage Ansible Modules
+# Version 7.0.0
 # Copyright (C) 2019-2022 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
@@ -102,7 +102,7 @@ options:
 requirements:
     - "python >= 3.8.6"
 notes:
-    - Run this module from a system that has direct access to DellEMC OpenManage Enterprise or OpenManage Enterprise Modular.
+    - Run this module from a system that has direct access to Dell OpenManage Enterprise or OpenManage Enterprise Modular.
     - I(device_group_names) option is not applicable for OpenManage Enterprise Modular.
     - This module supports C(check_mode).
 '''