+++ plugins/modules/ome_server_interface_profiles.py
@@ -2,8 +2,8 @@
 # -*- coding: utf-8 -*-
 
 #
-# Dell EMC OpenManage Ansible Modules
-# Version 5.1.0
+# Dell OpenManage Ansible Modules
+# Version 7.0.0
 # Copyright (C) 2022 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
@@ -103,7 +103,7 @@ requirements:
 author: "Jagadeesh N V (@jagadeeshnv)"
 notes:
     - This module supports C(check_mode).
-    - Run this module from a system that has direct access to Dell EMC OpenManage Enterprise Modular.
+    - Run this module from a system that has direct access to Dell OpenManage Enterprise Modular.
 '''
 
 EXAMPLES = r'''