+++ tests/unit/plugins/modules/test_ome_configuration_compliance_info.py
@@ -3,7 +3,7 @@
 #
 # Dell OpenManage Ansible Modules
 # Version 6.1.0
-# Copyright (C) 2021 Dell Inc. or its subsidiaries. All Rights Reserved.
+# Copyright (C) 2021-2022 Dell Inc. or its subsidiaries. All Rights Reserved.
 
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 #